Release C foundations. New products (Geosat, RADS). RADS service. Export ASCII. Remote Counter.
isardSAT/Geosat
- Geosat GDR product: new classes, aliases and other changes for BRAT to read the Geosat GDR netCDF products
- merge with DME work: updated geosat constructor signature and instantiated product class

RADS
- product identification
- product  class
- reformulation of brathl CheckFiles
- creation of new rads workspace data classes and specialization of existing ones
- persistence (I/O) of rads workspace classes (incomplete)
- GUI: RADS datasets tab, specialized tree widget (incomplete)

RADS Service
- implemented RADS service framework (windows)
- rsync corrections for mac
- work on configuration files reading and writing
- information sharing by brat and service

Edit Export ASCII Dialog
- implemented

brathl generic work
- changes in CProduct hierarchy and related classes to allow full data processing only on demand
- impact changes of previous item in filters, workspace related classes and GUI

GUI  generic work
- corrections in plot tracks
- progress dialog improvements
- improvements: string/vector algorithms, progress dialog, main map/window layout
- Qt5 bug correction in drag/drop

Remote counter
- tentative code using web engine: linux hacks using precompiled webengine
- remote counter draft windows and mac, working with unlisted
- web engine hacks removed
- simplified implementation without webengine

General
- corrections in release B to compile release C
- version pushed to 4.1.0
- new high level module common: extensive source files re-arrangement and impact changes in all development tools settings
- updated CMakeLists.txt with all changes in this commit
- new mac libs
- updated gitignore
- cppdend projects
- removed windows specific test code

+++ libbrathl/BratAlgoFilterKernel.h
@@ -22,7 +22,7 @@
 
 #include <string>
 
-#include "new-gui/Common/tools/brathl_error.h" 
+#include "common/tools/brathl_error.h"
 #include "brathl.h" 
 
 #include "Tools.h" 
@@ -95,7 +95,7 @@ protected:
   //virtual void PrepareDataWindow2D(double* dataValue, uint32_t nbValues);
 
   virtual void PrepareDataWindow2DSymmetrically(CExpressionValue& exprValue, uint32_t shiftSymmetry = 0);
-  virtual void PrepareDataWindow2DSymmetrically(double* dataValue, uint32_t nbValues, uint32_t shiftSymmetry = 0);
+  virtual void PrepareDataWindow2DSymmetrically(double* dataValue, size_t nbValues, uint32_t shiftSymmetry = 0);
 
   void SetParamValueExtrapolate(CVectorBratAlgorithmParam& args, uint32_t paramIndex);
